requirements.txt
apache-airflow-providers-amazon
pandas
openpyxl
boto3
pyarrow